kafka_consumer/requirements.txt
kafka-python==2.0.2
redis==5.2.1
nltk==3.9.1
celery==5.4.0
gspread==6.1.4
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1